+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/activity/SessionCommitActivity.kt
@@ -204,6 +204,7 @@ internal abstract class SessionCommitActivity<F : Failure> protected constructor
 				if (!isOnActivityResultCalled) {
 					abortSession()
 				}
+				finish()
 			}
 		}
 	}